requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
backports.zoneinfo==0.2.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==2.4.0
channels-redis==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.0
daphne==2.5.0
Django==3.0.5
google-auth==2.16.0
hyperlink==21.0.0
idna==3.4
include==0.2.2
incremental==22.10.0
kubernetes==17.17.0
msgpack==1.0.4
mysqlclient==2.1.1
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
redis==4.4.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
websocket-client==1.5.1
zope.interface==5.5.2